Make MGTransferMatrixFree/MGTransferBlockMatrixFree alias
