+++ include/deal.II/multigrid/mg_transfer_global_coarsening.templates.h
@@ -49,7 +49,6 @@
 
 #include <deal.II/multigrid/mg_tools.h>
 #include <deal.II/multigrid/mg_transfer_global_coarsening.h>
-#include <deal.II/multigrid/mg_transfer_matrix_free.templates.h>
 
 #include <boost/algorithm/string/join.hpp>
 
@@ -4745,6 +4744,64 @@ MGTransferMF<dim, Number, MemorySpace>::clear()
   internal_transfer.clear();
   transfer.clear();
   external_partitioners.clear();
+}
+
+template <int dim, typename Number, typename TransferType>
+void
+MGTransferBlockMatrixFreeBase<dim, Number, TransferType>::prolongate(
+  const unsigned int                                     to_level,
+  LinearAlgebra::distributed::BlockVector<Number>       &dst,
+  const LinearAlgebra::distributed::BlockVector<Number> &src) const
+{
+  const unsigned int n_blocks = src.n_blocks();
+  AssertDimension(dst.n_blocks(), n_blocks);
+
+  for (unsigned int b = 0; b < n_blocks; ++b)
+    {
+      const unsigned int data_block = same_for_all ? 0 : b;
+      get_matrix_free_transfer(data_block)
+        .prolongate(to_level, dst.block(b), src.block(b));
+    }
+}
+
+
+
+template <int dim, typename Number, typename TransferType>
+void
+MGTransferBlockMatrixFreeBase<dim, Number, TransferType>::prolongate_and_add(
+  const unsigned int                                     to_level,
+  LinearAlgebra::distributed::BlockVector<Number>       &dst,
+  const LinearAlgebra::distributed::BlockVector<Number> &src) const
+{
+  const unsigned int n_blocks = src.n_blocks();
+  AssertDimension(dst.n_blocks(), n_blocks);
+
+  for (unsigned int b = 0; b < n_blocks; ++b)
+    {
+      const unsigned int data_block = same_for_all ? 0 : b;
+      get_matrix_free_transfer(data_block)
+        .prolongate_and_add(to_level, dst.block(b), src.block(b));
+    }
+}
+
+
+
+template <int dim, typename Number, typename TransferType>
+void
+MGTransferBlockMatrixFreeBase<dim, Number, TransferType>::restrict_and_add(
+  const unsigned int                                     from_level,
+  LinearAlgebra::distributed::BlockVector<Number>       &dst,
+  const LinearAlgebra::distributed::BlockVector<Number> &src) const
+{
+  const unsigned int n_blocks = src.n_blocks();
+  AssertDimension(dst.n_blocks(), n_blocks);
+
+  for (unsigned int b = 0; b < n_blocks; ++b)
+    {
+      const unsigned int data_block = same_for_all ? 0 : b;
+      get_matrix_free_transfer(data_block)
+        .restrict_and_add(from_level, dst.block(b), src.block(b));
+    }
 }
