+++ source/multigrid/mg_transfer_matrix_free.inst.in
@@ -16,10 +16,4 @@
 
 for (deal_II_dimension : DIMENSIONS; S1 : REAL_SCALARS)
   {
-    template class MGTransferMatrixFree<deal_II_dimension, S1>;
-    template class MGTransferBlockMatrixFreeBase<
-      deal_II_dimension,
-      S1,
-      MGTransferMatrixFree<deal_II_dimension, S1>>;
-    template class MGTransferBlockMatrixFree<deal_II_dimension, S1>;
   }
